+++ package.json
@@ -15,38 +15,38 @@
 		"common": "workspace:*",
 		"kysely": "^0.27.5",
 		"kysely-d1": "^0.3.0",
-		"loro-crdt": "^1.4.2",
+		"loro-crdt": "^1.4.5",
 		"partysocket": "^1.0.3"
 	},
 	"devDependencies": {
-		"@cloudflare/workers-types": "^4.20250204.0",
-		"@eslint/js": "^9.20.0",
-		"@sveltejs/adapter-cloudflare": "^5.0.2",
-		"@sveltejs/kit": "^2.17.1",
+		"@cloudflare/workers-types": "^4.20250224.0",
+		"@eslint/js": "^9.21.0",
+		"@sveltejs/adapter-cloudflare": "^5.0.3",
+		"@sveltejs/kit": "^2.17.3",
 		"@sveltejs/vite-plugin-svelte": "^5.0.3",
-		"@tailwindcss/vite": "^4.0.6",
+		"@tailwindcss/vite": "^4.0.9",
 		"@types/better-sqlite3": "^7.6.12",
 		"better-sqlite3": "^11.8.1",
 		"dotenv": "^16.4.7",
-		"eslint": "^9.20.1",
-		"eslint-config-prettier": "^10.0.1",
+		"eslint": "^9.21.0",
+		"eslint-config-prettier": "^10.0.2",
 		"eslint-plugin-prettier": "^5.2.3",
 		"eslint-plugin-svelte": "^2.46.1",
-		"globals": "^15.14.0",
-		"prettier": "^3.5.0",
+		"globals": "^15.15.0",
+		"prettier": "^3.5.2",
 		"prettier-plugin-svelte": "^3.3.3",
 		"prettier-plugin-tailwindcss": "^0.6.11",
-		"prisma": "^6.3.1",
+		"prisma": "^6.4.1",
 		"prisma-kysely": "^1.8.0",
-		"svelte": "^5.19.10",
+		"svelte": "^5.20.5",
 		"svelte-check": "^4.1.4",
-		"tailwindcss": "^4.0.6",
+		"tailwindcss": "^4.0.9",
 		"typescript": "~5.7.3",
-		"typescript-eslint": "^8.24.0",
-		"vite": "^6.1.0",
+		"typescript-eslint": "^8.25.0",
+		"vite": "^6.2.0",
 		"vite-plugin-top-level-await": "^1.5.0",
 		"vite-plugin-wasm": "^3.4.1",
-		"wrangler": "^3.108.1"
+		"wrangler": "^3.111.0"
 	},
 	"packageManager": null
 }
