+++ src/lib/nodes/ListPickerNode.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import { Handle, Position, type NodeProps, type Node } from '@xyflow/svelte';
-  import { getFlowDoc} from '../LoroDoc.svelte';
+  import { getFlowDoc} from './LoroDoc.svelte';
   type ListPickerNode = Node<{
     label: string;
     options: string[];
