+++ src/lib/nodes/TextListNode2.svelte
@@ -1,25 +1,24 @@
 <script lang="ts">
   import { Handle, Position, useSvelteFlow, type NodeProps, type Node } from '@xyflow/svelte';
-  import type { MediaItem } from './node-types.svelte';
-  import { updateNodeDataAndSync } from '../LoroDoc.svelte';
-	import { getFlowDoc } from '../LoroDoc.svelte';
+  import type { TextItem } from './node-types.svelte';
+  import { getFlowDoc, updateNodeDataAndSync } from './LoroDoc.svelte';
 
-  type MediaListNodeType = Node<{
+  type TextListNodeType = Node<{
     label: string;
-    items: MediaItem[];
-  }, 'mediaList'>;
+    items: TextItem[];
+  }, 'textList'>;
 
-  let { id, data, positionAbsoluteX, positionAbsoluteY }: NodeProps<MediaListNodeType> = $props();
+  let { id, data, positionAbsoluteX, positionAbsoluteY }: NodeProps<TextListNodeType> = $props();
   const { updateNodeData } = useSvelteFlow();
   const flowDoc = getFlowDoc();
   function addItem() {
-    const newItems = [...data.items, { id: crypto.randomUUID(), text: '', imageUrl: '' }];
+    const newItems = [...data.items, { id: crypto.randomUUID(), text: '' }];
     updateNodeDataAndSync(
       flowDoc,
       id,
       { items: newItems },
       updateNodeData,
-      'mediaList',
+      'textList',
       { x: positionAbsoluteX, y: positionAbsoluteY }
     );
   }
@@ -31,22 +30,22 @@
       id,
       { items: newItems },
       updateNodeData,
-      'mediaList',
+      'textList',
       { x: positionAbsoluteX, y: positionAbsoluteY }
     );
   }
 
-  function updateItem(itemId: string, updates: Partial<MediaItem>) {
+  function updateItemText(itemId: string, newText: string) {
     const newItems = data.items.map(item => 
-      item.id === itemId ? { ...item, ...updates } : item
+      item.id === itemId ? { ...item, text: newText } : item
     );
     updateNodeDataAndSync(
       flowDoc,
       id,
       { items: newItems },
       updateNodeData,
-      'mediaList',
-      { x: positionAbsoluteX, y: positionAbsoluteY }
+      'textList',
+      // { x: positionAbsoluteX, y: positionAbsoluteY }
     );
   }
 
@@ -57,28 +56,13 @@
       id,
       { label: input.value },
       updateNodeData,
-      'mediaList',
-      { x: positionAbsoluteX, y: positionAbsoluteY }
+      'textList',
+      // { x: positionAbsoluteX, y: positionAbsoluteY }
     );
-  }
-
-  async function handleImageUpload(event: Event, itemId: string) {
-    const input = event.target as HTMLInputElement;
-    const file = input.files?.[0];
-    
-    if (file) {
-      // Convert the file to a data URL
-      const reader = new FileReader();
-      reader.onload = (e) => {
-        const imageUrl = e.target?.result as string;
-        updateItem(itemId, { imageUrl });
-      };
-      reader.readAsDataURL(file);
-    }
   }
 </script>
 
-<div class="media-list-node">
+<div class="text-list-node">
   <Handle type="target" position={Position.Left} />
   
   <div class="title">
@@ -89,38 +73,13 @@
   <div class="items">
     {#each data.items as item (item.id)}
       <div class="item">
-        <div class="media-container">
-          {#if item.imageUrl}
-            <img
-              src={item.imageUrl}
-              alt="Item"
-              class="preview-image"
-            />
-          {:else}
-            <div class="image-placeholder">
-              <label for="image-{item.id}" class="upload-label nodrag">
-                Upload Image
-              </label>
-            </div>
-          {/if}
-          <input
-            type="file"
-            id="image-{item.id}"
-            accept="image/*"
-            class="nodrag"
-            onchange={(e) => handleImageUpload(e, item.id)}
-            style="display: none;"
-          />
-        </div>
-        
         <input
           type="text"
           value={item.text}
-          oninput={(e) => updateItem(item.id, { text: e.currentTarget.value })}
+          oninput={(e) => updateItemText(item.id, e.currentTarget.value)}
           placeholder="Enter text..."
           class="nodrag"
         />
-        
         <button 
           class="remove-btn nodrag" 
           onclick={() => removeItem(item.id)}
@@ -128,7 +87,6 @@
         >
           ×
         </button>
-        
         <Handle
           type="source"
           position={Position.Right}
@@ -141,12 +99,12 @@
 </div>
 
 <style>
-  .media-list-node {
+  .text-list-node {
     padding: 10px;
     border-radius: 5px;
     background: white;
     border: 1px solid #ddd;
-    min-width: 300px;
+    min-width: 200px;
   }
 
   .title {
@@ -161,56 +119,17 @@
   .items {
     display: flex;
     flex-direction: column;
-    gap: 12px;
+    gap: 8px;
   }
 
   .item {
     display: flex;
     align-items: center;
-    gap: 8px;
     position: relative;
     padding-right: 20px;
   }
 
-  .media-container {
-    width: 60px;
-    height: 60px;
-    border: 1px dashed #ccc;
-    border-radius: 4px;
-    overflow: hidden;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-  }
-
-  .preview-image {
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-  }
-
-  .image-placeholder {
-    width: 100%;
-    height: 100%;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    background: #f5f5f5;
-  }
-
-  .upload-label {
-    font-size: 10px;
-    color: #666;
-    cursor: pointer;
-    text-align: center;
-    padding: 4px;
-  }
-
-  .upload-label:hover {
-    color: #333;
-  }
-
-  input[type="text"] {
+  input {
     flex: 1;
     padding: 6px;
     border: 1px solid #ddd;
@@ -218,7 +137,7 @@
     font-size: 12px;
   }
 
-  input[type="text"]:focus {
+  input:focus {
     outline: none;
     border-color: #aaa;
   }
@@ -250,4 +169,4 @@
   .remove-btn:hover {
     color: #ff0000;
   }
-</style> 
+</style>
