+++ src/lib/nodes/TextNode.svelte
