+++ worker/worker-configuration.d.ts
@@ -1,5 +1,7 @@
 // Generated by Wrangler by running `wrangler types`
 
 interface Env {
-	SOCKET_OBJECT: DurableObjectNamespace<import('./src/worker').SocketObject>;
+	FLOW_KV: KVNamespace;
+	SOCKET_OBJECT: DurableObjectNamespace<import("./src/worker").SocketObject>;
+	FLOW_BUCKET: R2Bucket;
 }
